agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.2",
    "firebase-admin>=7.1.0",
    "google-adk>=1.18.0",
    "google-cloud-aiplatform>=1.127.0",
    "google-cloud-firestore>=2.21.0",
    "google-cloud-storage>=3.5.0",
    "google-genai>=1.50.1",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "tenacity>=9.1.2",
    "uvicorn>=0.38.0",
    "websockets>=15.0.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.5",
]

[tool.ruff]
extend = "../pyproject.toml"
